Revert "Mess around with async_fn_in_trait"
This reverts commit 73d72cde985e1a42e8940f14fafe276a4aa0d129.

+++ src/bufpool/agg.rs
@@ -857,7 +857,7 @@ mod tests {
                 assert_eq!(len, block_size as usize);
                 slice.fill(2);
                 buf.advance(len as _);
-                assert_eq!(buf.len(), block_size * 2);
+                assert_eq!(buf.len(), block_size as u32 * 2);
             }
 
             {
@@ -970,7 +970,7 @@ mod tests {
         let mut pending = &input[..];
 
         loop {
-            let slice = buf.read().slice(0..buf.read().len());
+            let slice = buf.read().slice(0..buf.read().len() as u32);
             let (rest, version) = match parse(slice) {
                 Ok(t) => t,
                 Err(e) => {
